Revert "Merge NextStructure -> develop"
