+++ Client/EndPoints.cs
@@ -15,7 +15,6 @@ namespace SharpChatwork
         public static readonly Uri Contacts =                               new Uri($@"https://api.chatwork.com/v2/contacts");
         public static readonly Uri IncomingRequests =                       new Uri($@"https://api.chatwork.com/v2/incoming_requests");
 
-        public static Uri RoomOf(long roomId) =>                            new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}");
         public static Uri RoomMember(long roomId) =>                        new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}/members");
         public static Uri RoomMessages(long roomId) =>                      new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}/messages");
         public static Uri RoomMessagesOf(long roomId,long messageId) =>     new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}/messages/{messageId.ToString()}");
@@ -25,7 +24,6 @@ namespace SharpChatwork
         public static Uri RoomTasksOf(long roomId,long taskId) =>           new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}/tasks/{taskId.ToString()}");
         public static Uri RoomFiles(long roomId) =>                         new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}/files");
         public static Uri RoomFilesOf(long roomId, long fileId) =>          new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}/files/{fileId.ToString()}");
-        public static Uri RoomLink(long roomId) =>                          new Uri($@"https://api.chatwork.com/v2/rooms/{roomId.ToString()}/link");
         public static Uri IncomingRequestsOf(long requestId) =>             new Uri($@"https://api.chatwork.com/v2/incoming_requests/{requestId.ToString()}");
     }
 }
