--- Client/IQuery/IMeQuery.cs
@@ -1,15 +1,0 @@
-﻿using SharpChatwork.Query.Types;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SharpChatwork.Query
-{
-	public interface IMeQuery
-	{
-		ValueTask<Status> GetMyStatusAsync();
-		ValueTask<List<UserTask>> GetMyTasksAsync();
-		ValueTask<User> GetUserAsync();
-	}
-}
