--- Client/QueryResult/ElementId.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SharpChatwork.Query.Types
-{
-    public class ElementId
-    {
-        public long id { get; set; }
-    }
-}
