--- Client/QueryResult/Ids/MessageId.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SharpChatwork.Query.Types.Ids
-{
-    public class MessageId : ElementId
-    {
-        public long message_id
-        {
-            get => this.id;
-            set => this.id = value;
-        }
-    }
-}
