--- Client/RoomIconPreset.cs
@@ -1,44 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SharpChatwork
-{
-    public enum RoomIconPreset
-    {
-        [EnumAlias("group")]
-        Group,
-        [EnumAlias("check")]
-        Check,
-        [EnumAlias("document")]
-        Document,
-        [EnumAlias("meeting")]
-        Meeting, 
-        [EnumAlias("event")]
-        Event,
-        [EnumAlias("project")]
-        Project,
-        [EnumAlias("business")]
-        Business,
-        [EnumAlias("study")]
-        Study,
-        [EnumAlias("security")]
-        Security,
-        [EnumAlias("star")]
-        Star,
-        [EnumAlias("idea")]
-        Idea,
-        [EnumAlias("heart")]
-        Heart,
-        [EnumAlias("magcup")]
-        Magcup,
-        [EnumAlias("beer")]
-        Beer,
-        [EnumAlias("music")]
-        Music,
-        [EnumAlias("sports")]
-        Sports,
-        [EnumAlias("travel")]
-        Travel
-    }
-}
