+++ Program.cs
@@ -7,8 +7,6 @@ namespace SharpChatwork
         static void Main(string[] args)
         {
             Console.WriteLine("Hello World!");
-			SharpChatwork.IChatworkClient e = null;
-			e.room.message.SendAsync(0, "", true);
-		}
+        }
     }
 }
