+++ Query/IChatworkQueryResult.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SharpChatwork.Query
+{
+    public interface IChatworkQueryResult
+    {
+        ChatworkQueryInfo info { get; set; }
+    }
+}
