+++ Query/Types/Contact.cs
