+++ Query/Types/IncomingRequest.cs
