+++ Query/Types/Room.cs
