+++ Query/Types/UserTask.cs
@@ -1,13 +1,5 @@
 ﻿namespace SharpChatwork.Query.Types
 {
-    public enum TaskStateType
-    {
-        [EnumAlias("open")]
-        Open,
-        [EnumAlias("done")]
-        Done
-    }
-
     public class UserTask
     {
         public int task_id { get; set; }
